comments and convertion vec4->SDL_FRect

+++ Source/Enviroment/splitView.cpp
@@ -4,7 +4,7 @@ void SplitView::Render(const vec4<float>& area)
 {
 	// area.x = width
 	// area.y = height
-	int dividerY = (int)(area.y * ratio);
+	float dividerY = (area.y * ratio);
 
 	// Top panel
 	vec4<float> topRect = { 0, 0, area.x, dividerY };
